Ignore multiple possible gene preds for one contig
Pogigwasc may predict several equally scoring annotations for a single
contig. In that case, it reports 'ambiguous' in the log file and reports
all the annotations in the GFF output. Each block of annotations is
sorted by coordinate within the block, and blocks for the same contig
are output one after the other in the GFF with no gaps.

Since they are all equally well scoring we simply take the first one we
encounter and ignore the rest. It may be desirable in the future to
parse them and find the specific features that differ (usually most of
the features are predicted the same way between blocks and only a few
features are alternatively predicted). However given that only a few
features are affected, and these often fall in contigs with
low-complexity repeat regions, decide not to spend the effort for now.